requirements.txt
asgiref==3.8.1
Django==4.1.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
PyJWT==2.6.0